Move Check enum to separate file
